requirements/local.txt
-r base.txt # includes the base.txt requirements file

coverage
django-debug-toolbar
